refactor: Price Value Object를 공통 패키지로 리팩터링하라
