+++ src/main/java/kitchenpos/common/domain/vo/exception/InvalidPriceException.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.menu.tobe.domain.vo.exception;
+package kitchenpos.common.domain.vo.exception;
 
 public class InvalidPriceException extends IllegalArgumentException {
     private static final String MESSAGE = "가격은 0원 이상이여야 합니다.";
