+++ src/test/java/kitchenpos/menus/menu/tobe/domain/MenuTest.java
@@ -3,8 +3,8 @@ package kitchenpos.menus.menu.tobe.domain;
 import kitchenpos.common.domain.FakeProfanity;
 import kitchenpos.common.domain.Profanity;
 import kitchenpos.common.domain.vo.DisplayedName;
+import kitchenpos.common.domain.vo.Price;
 import kitchenpos.menus.menu.tobe.domain.exception.InvalidMenuException;
-import kitchenpos.menus.menu.tobe.domain.vo.Price;
 import kitchenpos.menus.menu.tobe.domain.vo.Quantity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
